pull AdminGrid out of the admin controller
